reading from file almost done

+++ Task3/src/main/java/by/epam/buhai/airline/entity/Airplane.java
@@ -17,7 +17,7 @@ public class Airplane extends PassengerPlane implements Serializable {
     }
 
     public Airplane(String name, Manufacturers manufacturer, int crew, int maxSpeedKmPerHour,
-                    double rangeKm, double fuelConsumptionLitersPerHour, int seatingCapacity,
+                    int rangeKm, int fuelConsumptionLitersPerHour, int seatingCapacity,
                     AirplaneTypes airplaneType, AirplaneBodyTypes airplaneBodyType,
                     AirplaneClassCabinsNumber airplaneClassCabinsNumber) {
